+++ FallLady.Mood.Domain/Domain/Courses/Course.cs
@@ -23,7 +23,7 @@ namespace FallLady.Mood.Domain.Domain.Courses
         }
         public Course(string title,
                       CourseTypeEnum courseType,
-                      decimal price,
+                      Int64 price,
                       string description,
                       string? licenseKey,
                       string fileName,
@@ -68,7 +68,7 @@ namespace FallLady.Mood.Domain.Domain.Courses
         #region Properties
         public string Title { get; private set; }
         public CourseTypeEnum CourseType { get; private set; }
-        public decimal Price { get; private set; }
+        public Int64 Price { get; private set; }
         public string Description { get; private set; }
         public string? LicenseKey { get; private set; }
         public string FileName { get; private set; }
@@ -97,7 +97,7 @@ namespace FallLady.Mood.Domain.Domain.Courses
 
         public Course Update(string title,
                              CourseTypeEnum courseType,
-                             decimal price,
+                             Int64 price,
                              string description,
                              string? licenseKey,
                              string fileName,
